ApiEnums  and ApiConstants moved to api folder
