+++ WhaleSpotting/react-app/src/api/ApiEnums.ts
